--- src/main/java/com/example/frankfurter/config/FrankfurterClientProperties.java
@@ -1,18 +1,0 @@
-package com.example.frankfurter.config;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-@ConfigurationProperties(prefix = "frankfurter")
-public class FrankfurterClientProperties {
-
-    private String baseUrl;
-
-    public String getBaseUrl() {
-        return baseUrl;
-    }
-
-    public void setBaseUrl(String baseUrl) {
-        this.baseUrl = baseUrl;
-    }
-
-}
