--- src/main/java/com/example/frankfurter/config/GithubProperties.java
@@ -1,19 +1,0 @@
-package com.example.frankfurter.config;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.stereotype.Component;
-
-@Component
-@ConfigurationProperties(prefix = "github")
-public class GithubProperties {
-
-    private String username;
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username.toLowerCase();
-    }
-}
