--- src/main/java/com/example/frankfurter/dto/FrankfurterLatestResponse.java
@@ -1,34 +1,0 @@
-package com.example.frankfurter.dto;
-
-import java.util.Map;
-
-public class FrankfurterLatestResponse {
-
-    private String base;
-    private String date;
-    private Map<String, Double> rates;
-
-    public String getBase() {
-        return base;
-    }
-
-    public void setBase(String base) {
-        this.base = base;
-    }
-
-    public String getDate() {
-        return date;
-    }
-
-    public void setDate(String date) {
-        this.date = date;
-    }
-
-    public Map<String, Double> getRates() {
-        return rates;
-    }
-
-    public void setRates(Map<String, Double> rates) {
-        this.rates = rates;
-    }
-}
