--- src/main/java/com/example/frankfurter/dto/HistoricalIdrUsdDto.java
@@ -1,20 +1,0 @@
-package com.example.frankfurter.dto;
-
-public class HistoricalIdrUsdDto {
-
-    private String date;
-    private double usdRate;
-
-    public HistoricalIdrUsdDto(String date, double usdRate) {
-        this.date = date;
-        this.usdRate = usdRate;
-    }
-
-    public String getDate() {
-        return date;
-    }
-
-    public double getUsdRate() {
-        return usdRate;
-    }
-}
