--- src/main/java/com/example/frankfurter/exception/DataNotLoadedException.java
@@ -1,7 +1,0 @@
-package com.example.frankfurter.exception;
-
-public class DataNotLoadedException extends RuntimeException {
-    public DataNotLoadedException(String message) {
-        super(message);
-    }
-}
