--- src/main/java/com/example/frankfurter/exception/ExternalApiException.java
@@ -1,7 +1,0 @@
-package com.example.frankfurter.exception;
-
-public class ExternalApiException extends RuntimeException {
-    public ExternalApiException(String message) {
-        super(message);
-    }
-}
