--- src/main/java/com/example/frankfurter/exception/GlobalExceptionHandler.java
@@ -1,43 +1,0 @@
-package com.example.frankfurter.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-import java.time.Instant;
-import java.util.Map;
-
-@RestControllerAdvice
-public class GlobalExceptionHandler {
-
-    @ExceptionHandler(UnknownResourceTypeException.class)
-    public ResponseEntity<Map<String, Object>> handleUnknownResource(UnknownResourceTypeException ex) {
-        return build(HttpStatus.BAD_REQUEST, ex.getMessage());
-    }
-
-    @ExceptionHandler(DataNotLoadedException.class)
-    public ResponseEntity<Map<String, Object>> handleDataNotLoaded(DataNotLoadedException ex) {
-        return build(HttpStatus.SERVICE_UNAVAILABLE, ex.getMessage());
-    }
-
-    @ExceptionHandler(ExternalApiException.class)
-    public ResponseEntity<Map<String, Object>> handleExternalApi(ExternalApiException ex) {
-        return build(HttpStatus.BAD_GATEWAY, ex.getMessage());
-    }
-
-    @ExceptionHandler(Exception.class)
-    public ResponseEntity<Map<String, Object>> handleGeneric(Exception ex) {
-        return build(HttpStatus.INTERNAL_SERVER_ERROR, ex.getMessage());
-    }
-
-    private ResponseEntity<Map<String, Object>> build(HttpStatus status, String message) {
-        Map<String, Object> body = Map.of(
-                "timestamp", Instant.now().toString(),
-                "status", status.value(),
-                "error", status.getReasonPhrase(),
-                "message", message
-        );
-        return ResponseEntity.status(status).body(body);
-    }
-}
