--- src/main/java/com/example/frankfurter/exception/UnknownResourceTypeException.java
@@ -1,7 +1,0 @@
-package com.example.frankfurter.exception;
-
-public class UnknownResourceTypeException extends RuntimeException {
-    public UnknownResourceTypeException(String resourceType) {
-        super("Unknown resource type: " + resourceType);
-    }
-}
