+++ src/main/java/com/thasya/frankfurter/config/FrankfurterClientProperties.java
@@ -1,0 +1,18 @@
+package com.thasya.frankfurter.config;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+@ConfigurationProperties(prefix = "frankfurter")
+public class FrankfurterClientProperties {
+
+    private String baseUrl;
+
+    public String getBaseUrl() {
+        return baseUrl;
+    }
+
+    public void setBaseUrl(String baseUrl) {
+        this.baseUrl = baseUrl;
+    }
+
+}
