+++ src/main/java/com/thasya/frankfurter/dto/CurrencyDto.java
@@ -1,0 +1,20 @@
+package com.thasya.frankfurter.dto;
+
+public class CurrencyDto {
+
+    private String code;
+    private String description;
+
+    public CurrencyDto(String code, String description) {
+        this.code = code;
+        this.description = description;
+    }
+
+    public String getCode() {
+        return code;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+}
