+++ src/main/java/com/thasya/frankfurter/dto/FrankfurterTimeseriesResponse.java
@@ -1,0 +1,43 @@
+package com.thasya.frankfurter.dto;
+
+import java.util.Map;
+
+public class FrankfurterTimeseriesResponse {
+
+    private String start_date;
+    private String end_date;
+    private String base;
+    private Map<String, Map<String, Double>> rates;
+
+    public String getStart_date() {
+        return start_date;
+    }
+
+    public void setStart_date(String start_date) {
+        this.start_date = start_date;
+    }
+
+    public String getEnd_date() {
+        return end_date;
+    }
+
+    public void setEnd_date(String end_date) {
+        this.end_date = end_date;
+    }
+
+    public String getBase() {
+        return base;
+    }
+
+    public void setBase(String base) {
+        this.base = base;
+    }
+
+    public Map<String, Map<String, Double>> getRates() {
+        return rates;
+    }
+
+    public void setRates(Map<String, Map<String, Double>> rates) {
+        this.rates = rates;
+    }
+}
