+++ src/main/java/com/thasya/frankfurter/dto/HistoricalIdrUsdDto.java
@@ -1,0 +1,20 @@
+package com.thasya.frankfurter.dto;
+
+public class HistoricalIdrUsdDto {
+
+    private String date;
+    private double usdRate;
+
+    public HistoricalIdrUsdDto(String date, double usdRate) {
+        this.date = date;
+        this.usdRate = usdRate;
+    }
+
+    public String getDate() {
+        return date;
+    }
+
+    public double getUsdRate() {
+        return usdRate;
+    }
+}
