+++ src/main/java/com/thasya/frankfurter/exception/DataNotLoadedException.java
@@ -1,0 +1,7 @@
+package com.thasya.frankfurter.exception;
+
+public class DataNotLoadedException extends RuntimeException {
+    public DataNotLoadedException(String message) {
+        super(message);
+    }
+}
