+++ src/main/java/com/thasya/frankfurter/exception/ExternalApiException.java
@@ -1,0 +1,7 @@
+package com.thasya.frankfurter.exception;
+
+public class ExternalApiException extends RuntimeException {
+    public ExternalApiException(String message) {
+        super(message);
+    }
+}
