+++ src/main/java/com/thasya/frankfurter/exception/UnknownResourceTypeException.java
@@ -1,0 +1,7 @@
+package com.thasya.frankfurter.exception;
+
+public class UnknownResourceTypeException extends RuntimeException {
+    public UnknownResourceTypeException(String resourceType) {
+        super("Unknown resource type: " + resourceType);
+    }
+}
