--- src/test/java/com/example/frankfurter/FrankfurterApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.frankfurter;
-
-import org.junit.jupiter.api.Test;
-
-class FrankfurterApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-
-
-}
