+++ src/test/java/com/thasya/frankfurter/FrankfurterApplicationTests.java
@@ -1,0 +1,11 @@
+package com.thasya.frankfurter;
+
+import org.junit.jupiter.api.Test;
+
+class FrankfurterApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
